+++ ReleaseNotes/release_notes.md
@@ -11,6 +11,8 @@
   1. add `TH(2,3)(F,D).bezier`, `TH(2,3)(F,D).bezier_fast`, `TH(2,3)(F,D).bezier_fill` methods for 2&3D-histogram parameterisations based on `Bernstein2D::fill` and `Bernstein3D::fill` methods 
   1. add `TH(2,3)(F,D).bernstein`, `TH(2,3)(F,D).bernstein_fasr`, `TH(2,3)(F,D).bernstein_fill` methods for 2&3D-histogram parameterisations based on `Bernstein2D::fill` and `Bernstein3D::fill` methods 
   1. add `ostap.utils.utils.slow`  method for `slow`-iteration with delaye at each step
+  1. add `Ostap::Math::agm` for complex numbers 
+  1. add `Ostap::Math::agm` for `ValueWithError` objects 
  
 `
 ## Backward incompatible: